# packages/schemas/api/portal-openapi.yaml
openapi: "3.1.0"
info:
  title: Enterprise AI Asset Hub — Portal API
  version: "0.1.0"
  description: |
    Portal API (M02) — Asset Registry, Service Composer, Deployment Management.
    This is a draft contract. Implementation starts in Phase 1.
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /api/v1/asset-versions/lifecycle:
    get:
      tags: [assets]
      summary: P16 수명주기 목록 — every AssetVersion with its lifecycle state, replacement pointer, and active revocation (RELEASE_MANAGER/ADMIN/AUDITOR)
      operationId: list_asset_version_lifecycle
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetVersionLifecycleListResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/v1/audit-events:
    get:
      tags: [audit]
      summary: Search audit events (AUDITOR/ADMIN only)
      operationId: list_audit_events
      parameters:
        - name: resource_id
          in: query
          schema:
            type: string
        - name: event_type
          in: query
          schema:
            type: string
        - name: actor_id
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Audit events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditEventListResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/admin/settings:
    get:
      tags: [admin]
      summary: >
        P15 관리자 설정 — read-only "정책·구성 현황". Every section is either
        AVAILABLE (sourced live from its real config/policy file at request
        time, never a hand-maintained copy) or NOT_IMPLEMENTED (named reason
        + open-decisions.md id, never an empty form or a guessed value).
        ADMIN only. Never returns a live/original credential value — any
        config field that could hold one is redacted (see
        AdminSettingsOut.redaction_note on each section that reads a file).
      operationId: get_admin_settings
      responses:
        "200":
          description: Effective platform configuration snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettingsOut"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
components:
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, trace_id]
          properties:
            code:
              type: string
            message:
              type: string
            trace_id:
              type: string
            details:
              type: object
...
    AssetStatus:
      type: string
      enum:
        - DRAFT
        - VALIDATING
        - READY_FOR_REVIEW
        - IN_REVIEW
        - APPROVED
        - REJECTED
        - CHANGES_REQUESTED
        - DEPRECATED
        - RETIRED
        - SUSPENDED
...
    AssetVersionRevocation:
      type: object
      required: [id, version_id, reason, approver_id, effective_at, created_at, effective]
      properties:
        id:
          type: string
          format: uuid
        version_id:
          type: string
          format: uuid
        reason:
          type: string
        approver_id:
          type: string
        effective_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        trace_id:
          type: string
        effective:
          type: boolean
          description: Computed at read time — `effective_at <= now()`.
...
    AssetVersionLifecycleItem:
      type: object
      required: [id, asset_id, asset_name, asset_type, version, status]
      properties:
        id:
          type: string
          format: uuid
        asset_id:
          type: string
          format: uuid
        asset_name:
          type: string
        asset_type:
          type: string
        version:
          type: string
        status:
          $ref: "#/components/schemas/AssetStatus"
        approved_at:
          type: string
          format: date-time
          nullable: true
        deprecated_at:
          type: string
          format: date-time
          nullable: true
        retired_at:
          type: string
          format: date-time
          nullable: true
        replacement_version_id:
          type: string
          format: uuid
          nullable: true
        replacement_version_label:
          type: string
          nullable: true
        active_revocation:
          allOf:
            - $ref: "#/components/schemas/AssetVersionRevocation"
          nullable: true
    AssetVersionLifecycleListResponse:
      type: object
      required: [items, page, page_size, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AssetVersionLifecycleItem"
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
...
    AuditEvent:
      type: object
      required: [id, timestamp, event_type, actor_type, actor_id, resource_type, resource_id, result]
      properties:
        id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        event_type:
          type: string
        actor_type:
          type: string
        actor_id:
          type: string
        organization_id:
          type: string
        site_id:
          type: string
        resource_type:
          type: string
        resource_id:
          type: string
        resource_version:
          type: string
        trace_id:
          type: string
        run_id:
          type: string
        job_id:
          type: string
        result:
          type: string
          enum: [SUCCESS, DENIED, FAILED]
        policy_id:
          type: string
        metadata:
          type: object
          description: Sanitized summary only — never prompt text, full documents, or secrets.
    AuditEventListResponse:
      type: object
      required: [items, page, page_size, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AuditEvent"
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
...
    NotImplementedInfo:
      type: object
      description: >
        Recorded reason a P15 sub-area has no editor/data instead of an
        empty form or a guessed value (01-portal-and-distribution.md §2
        P15).
      required: [reason, decision_id]
      properties:
        reason: { type: string }
        decision_id:
          type: string
          description: open-decisions.md id this gap is tracked under.
    RolePermissionRow:
      type: object
      required: [role, permissions]
      properties:
        role: { type: string }
        permissions:
          type: array
          items: { type: string }
    UserRoleMappingSection:
      type: object
      description: >
        The real Role→Permission matrix (security_policy.ROLE_PERMISSIONS,
        7 roles). There is no user directory in this PoC (no `users` table)
        — this section shows what each role *can do*, not who holds it.
      required: [status, source, roles, note]
      properties:
        status: { type: string, enum: [AVAILABLE] }
        source: { type: string }
        roles:
          type: array
          items:
            $ref: "#/components/schemas/RolePermissionRow"
        note: { type: string }
    OrganizationSiteSection:
      type: object
      required: [status, not_implemented]
      properties:
        status: { type: string, enum: [NOT_IMPLEMENTED] }
        not_implemented:
          $ref: "#/components/schemas/NotImplementedInfo"
    ModelAliasSetting:
      type: object
      required: [alias, provider, model_id, endpoint]
      properties:
        alias: { type: string }
        provider: { type: string }
        model_id: { type: string }
        endpoint:
          type: string
          description: Redacted to ***REDACTED*** if it looks credential-shaped (security_policy.redact_if_secret).
        max_context_tokens: { type: integer }
    McpServerAliasSetting:
      type: object
      required: [alias, endpoint]
      properties:
        alias: { type: string }
        endpoint: { type: string }
        allowed_tools:
          type: array
          items: { type: string }
    OfficeProfileSection:
      type: object
      description: services/agent-runtime/config/office-profile-default/office-profile.json, read fresh at request time.
      required: [status, source, name, version, org, sites]
      properties:
        status: { type: string, enum: [AVAILABLE] }
        source: { type: string }
        name: { type: string }
        version: { type: string }
        org: { type: string }
        sites:
          type: array
          items: { type: string }
        max_classification_allowed: { type: string }
        audit_retention_days: { type: integer }
    ModelEndpointAliasSection:
      type: object
      description: 허용 모델과 Endpoint Alias — same office-profile.json source as OfficeProfileSection.
      required: [status, source, model_aliases, mcp_servers]
      properties:
        status: { type: string, enum: [AVAILABLE] }
        source: { type: string }
        model_aliases:
          type: array
          items:
            $ref: "#/components/schemas/ModelAliasSetting"
        mcp_servers:
          type: array
          items:
            $ref: "#/components/schemas/McpServerAliasSetting"
    AssetSizeExtensionPolicySection:
      type: object
      description: >
        Two real, independent policies — knowledge_packager's build-time
        content scanner (extension/pattern based) and the Desktop
        offline-bundle installer's pre-extraction checks (byte-size based).
        Portal 자산 등록 자체(POST /api/v1/assets)에는 별도 크기/확장자 제한이
        없다는 사실도 함께 보고한다(추측하지 않음).
      required: [status, source, portal_upload_limit_note]
      properties:
        status: { type: string, enum: [AVAILABLE] }
        source: { type: string }
        portal_upload_limit_note: { type: string }
        knowledge_package_forbidden_filenames:
          type: array
          items: { type: string }
        knowledge_package_fail_on_fatal: { type: boolean }
        desktop_bundle_max_total_uncompressed_bytes: { type: integer }
        desktop_bundle_max_single_file_uncompressed_bytes: { type: integer }
        desktop_bundle_max_compression_ratio: { type: number }
        desktop_bundle_forbidden_archive_extensions:
          type: array
          items: { type: string }
        desktop_bundle_forbidden_executable_extensions:
          type: array
          items: { type: string }
        parse_error:
          type: string
          description: Set (in place of the desktop_bundle_*/knowledge_package_* fields above) if either source file is missing or no longer matches the expected shape — never a guessed value.
    ApprovalWorkflowSection:
      type: object
      description: security_policy's TECHNICAL→SECURITY→RELEASE chain + the ServiceVersion publish gate flag (D-063).
      required: [status, source, stage_chain, require_service_version_approval, require_service_version_approval_source]
      properties:
        status: { type: string, enum: [AVAILABLE] }
        source: { type: string }
        stage_chain:
          type: array
          items: { type: string }
        require_service_version_approval: { type: boolean }
        require_service_version_approval_source: { type: string }
    SecurityClassificationSubsection:
      type: object
      required: [status, source, classification_levels, acl_metadata_fields, allow_unknown_classification, allow_unknown_classification_caveat]
      properties:
        status: { type: string, enum: [AVAILABLE] }
        source: { type: string }
        classification_levels:
          type: array
          items: { type: string }
        acl_metadata_fields:
          type: array
          items: { type: string }
        allow_unknown_classification: { type: boolean }
        allow_unknown_classification_caveat: { type: string }
    RetentionSubsection:
      type: object
      required: [status, not_implemented]
      properties:
        status: { type: string, enum: [NOT_IMPLEMENTED] }
        not_implemented:
          $ref: "#/components/schemas/NotImplementedInfo"
    SecurityClassificationRetentionSection:
      type: object
      description: 보안등급과 보관기간 — one P15 bullet, two independent statuses (등급 real, 보관기간 미구현).
      required: [classification, retention]
      properties:
        classification:
          $ref: "#/components/schemas/SecurityClassificationSubsection"
        retention:
          $ref: "#/components/schemas/RetentionSubsection"
    PackageTrustSignatureSection:
      type: object
      required: [status, not_implemented]
      properties:
        status: { type: string, enum: [NOT_IMPLEMENTED] }
        not_implemented:
          $ref: "#/components/schemas/NotImplementedInfo"
    AdminSettingsOut:
      type: object
      required:
        - generated_at
        - trace_id
        - user_role_mapping
        - organization_site
        - office_profile
        - model_endpoint_alias
        - asset_size_extension_policy
        - approval_workflow
        - security_classification_retention
        - package_trust_signature
      properties:
        generated_at: { type: string, format: date-time }
        trace_id: { type: string }
        user_role_mapping:
          $ref: "#/components/schemas/UserRoleMappingSection"
        organization_site:
          $ref: "#/components/schemas/OrganizationSiteSection"
        office_profile:
          $ref: "#/components/schemas/OfficeProfileSection"
        model_endpoint_alias:
          $ref: "#/components/schemas/ModelEndpointAliasSection"
        asset_size_extension_policy:
          $ref: "#/components/schemas/AssetSizeExtensionPolicySection"
        approval_workflow:
          $ref: "#/components/schemas/ApprovalWorkflowSection"
        security_classification_retention:
          $ref: "#/components/schemas/SecurityClassificationRetentionSection"
        package_trust_signature:
          $ref: "#/components/schemas/PackageTrustSignatureSection"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []